+++ src/hooks/useAppSelector.ts
@@ -1,4 +1,4 @@
-import { RootState } from '@/redux/store';
+import { RootState } from '@/store/store';
 import { useSelector } from 'react-redux';
 
 const useAppSelector = useSelector.withTypes<RootState>();
