+++ src/layout/RootLayout/index.tsx
@@ -3,10 +3,10 @@ import * as S from './style';
 import { GUEST_UUID, LOGIN, MY_REGION } from '@/constants/localStorage/key';
 import { useMutation } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
-import { guestLogin } from '@/service/login';
+import { guestLogin } from '@/services/login';
 import useAppDispatch from '@/hooks/useAppDispatch';
 import A2hsSnackbar from './components/A2hsSnackbar';
-import { goelocationActions } from '@/redux/slice/geolocationSlice';
+import { goelocationActions } from '@/store/slice/geolocationSlice';
 import useGeolocation from '@/hooks/useGeolocation';
 import { storeAccessToken, storeUser } from '@/utils/auth';
 import useAppSelector from '@/hooks/useAppSelector';
