+++ src/pages/Feedback/index.tsx
@@ -4,7 +4,7 @@ import { Link, useNavigate } from 'react-router-dom';
 import { useState } from 'react';
 import CustomButton from '@/components/CustomMui/CustomButton';
 import { useMutation } from '@tanstack/react-query';
-import { submitFeedback } from '@/service/feedback';
+import { submitFeedback } from '@/services/feedback';
 import { useSnackbar } from '@/contexts/SnackbarProvider';
 import HeadHelmet from '@/components/HeadHelmet';
