+++ src/pages/UserGender/index.tsx
@@ -5,11 +5,11 @@ import FemaleImage from '@/assets/svg/gender/female.svg?react';
 import { useState } from 'react';
 import CustomButton from '@/components/CustomMui/CustomButton';
 import { useMutation } from '@tanstack/react-query';
-import { setUserGender } from '@/service/auth';
+import { setUserGender } from '@/services/auth';
 import useAppSelector from '@/hooks/useAppSelector';
 import { useNavigate } from 'react-router-dom';
 import useAppDispatch from '@/hooks/useAppDispatch';
-import { userActions } from '@/redux/slice/userSlice';
+import { userActions } from '@/store/slice/userSlice';
 import 바지 from '@/components/clothes/바지';
 import { Gender } from '@/types/user';
 import HeadHelmet from '@/components/HeadHelmet';
